1051 extract generation of *.html files to plugin PageSourceExtractor
