+++ src/main/resources/META-INF/defaultservices/com.codeborne.selenide.impl.PageSourceExtractor
@@ -1,0 +1,1 @@
+com.codeborne.selenide.impl.WebPageSourceExtractor
